requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
langchain==0.2.1
langchain-core==0.2.1
langchain-community==0.2.1
langchain-groq==0.1.4
langchain-openai==0.1.7
pyyaml==6.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
